pyproject.toml
[project]
name = "gemeni-pdf"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fitz>=0.0.1.dev2",
    "frontend>=0.0.3",
    "google-genai>=1.57.0",
    "pydantic>=2.12.5",
    "pymupdf>=1.26.7",
    "pypdf>=6.6.0",
    "python-dotenv>=1.2.1",
    "reportlab>=4.4.7",
    "tools>=1.0.17",
]
